Migrated ‘RJI’ test to proper `XCTest` test case
